Change RenderHost.ActualWidth/ActualHeight to int

+++ Source/HelixToolkit.SharpDX.Shared/Interface/IRenderHost.cs
@@ -150,14 +150,14 @@ namespace HelixToolkit.UWP
         /// <value>
         /// The actual height.
         /// </value>
-        double ActualHeight { get; }
+        int ActualHeight { get; }
         /// <summary>
         /// Gets the actual width.
         /// </summary>
         /// <value>
         /// The actual width.
         /// </value>
-        double ActualWidth { get; }
+        int ActualWidth { get; }
 
         /// <summary>
         /// Indicates if DPFCanvas busy on rendering.
@@ -258,7 +258,7 @@ namespace HelixToolkit.UWP
         /// </summary>
         /// <param name="width">The width.</param>
         /// <param name="height">The height.</param>
-        void StartD3D(double width, double height);
+        void StartD3D(int width, int height);
         /// <summary>
         /// Ends the d3 d.
         /// </summary>
@@ -284,7 +284,7 @@ namespace HelixToolkit.UWP
         /// </summary>
         /// <param name="width">The width.</param>
         /// <param name="height">The height.</param>
-        void Resize(double width, double height);
+        void Resize(int width, int height);
         /// <summary>
         /// Gets or sets a value indicating whether [show statistics].
         /// </summary>
